Change animation timing to seconds-based instead of ticks
Ticks per second was based on the frequency that requestAnimationFrame
was able to run. This is max'd out at a screen's refresh rate. On a new
macbook, this is 120fps, whereas on a phone it's 60. This was causing
quite different animation durations on different devices. Instead of
throttling the animation function, using seconds makes all animations
fps-agnostic.